Change test codes into proper directory (#260)
